Adding a simple base test class
Moving common code from the tests I just created and created a base class where they all extend from
