+++ src/test/java/org/jenkinsci/plugins/github_branch_source/GitSCMSourceBase.java
@@ -1,0 +1,91 @@
+package org.jenkinsci.plugins.github_branch_source;
+
+import com.github.tomakehurst.wiremock.common.FileSource;
+import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
+import com.github.tomakehurst.wiremock.extension.Parameters;
+import com.github.tomakehurst.wiremock.extension.ResponseTransformer;
+import com.github.tomakehurst.wiremock.http.Request;
+import com.github.tomakehurst.wiremock.http.Response;
+import com.github.tomakehurst.wiremock.junit.WireMockRule;
+import jenkins.scm.api.mixin.ChangeRequestCheckoutStrategy;
+import org.junit.Before;
+import org.junit.ClassRule;
+import org.junit.Rule;
+import org.jvnet.hudson.test.JenkinsRule;
+import org.kohsuke.github.GHRepository;
+import org.kohsuke.github.GitHub;
+
+import java.util.Arrays;
+import java.util.EnumSet;
+
+import static com.github.tomakehurst.wiremock.client.WireMock.*;
+import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
+
+public class GitSCMSourceBase{
+
+    /**
+     * All tests in this class only use Jenkins for the extensions
+     */
+    @ClassRule
+    public static JenkinsRule r = new JenkinsRule();
+
+    public static WireMockRuleFactory factory = new WireMockRuleFactory();
+
+    @Rule
+    public WireMockRule githubRaw = factory.getRule(WireMockConfiguration.options()
+            .dynamicPort()
+            .usingFilesUnderClasspath("raw")
+    );
+    @Rule
+    public WireMockRule githubApi = factory.getRule(WireMockConfiguration.options()
+            .dynamicPort()
+            .usingFilesUnderClasspath("api")
+            .extensions(
+                    new ResponseTransformer() {
+                        @Override
+                        public Response transform(Request request, Response response, FileSource files,
+                                                  Parameters parameters) {
+                            if ("application/json"
+                                    .equals(response.getHeaders().getContentTypeHeader().mimeTypePart())) {
+                                return Response.Builder.like(response)
+                                        .but()
+                                        .body(response.getBodyAsString()
+                                                .replace("https://api.github.com/",
+                                                        "http://localhost:" + githubApi.port() + "/")
+                                                .replace("https://raw.githubusercontent.com/",
+                                                        "http://localhost:" + githubRaw.port() + "/")
+                                        )
+                                        .build();
+                            }
+                            return response;
+                        }
+
+                        @Override
+                        public String getName() {
+                            return "url-rewrite";
+                        }
+
+                    })
+    );
+
+    GitHubSCMSource source;
+    GitHub github;
+    GHRepository repo;
+
+    @Before
+    public void prepareMockGitHub() throws Exception {
+        githubApi.stubFor(
+                get(urlMatching(".*")).atPriority(10).willReturn(aResponse().proxiedFrom("https://api.github.com/")));
+        githubRaw.stubFor(get(urlMatching(".*")).atPriority(10)
+                .willReturn(aResponse().proxiedFrom("https://raw.githubusercontent.com/")));
+
+        if (source.isConfiguredByUrl()) {
+            source = new GitHubSCMSource("cloudbeers", "yolo", "http://localhost:" + githubApi.port() + "/cloudbeers/yolo", true);
+        } else {
+            source.setApiUri("http://localhost:" + githubApi.port());
+        }
+        source.setTraits(Arrays.asList(new BranchDiscoveryTrait(true, true), new ForkPullRequestDiscoveryTrait(EnumSet.of(ChangeRequestCheckoutStrategy.MERGE), new ForkPullRequestDiscoveryTrait.TrustContributors())));
+        github = Connector.connect("http://localhost:" + githubApi.port(), null);
+        repo = github.getRepository("cloudbeers/yolo");
+    }
+}
