+++ src/test/java/org/jenkinsci/plugins/github_branch_source/GithubSCMSourceBranchesTest.java
@@ -1,117 +1,27 @@
 package org.jenkinsci.plugins.github_branch_source;
 
-import com.github.tomakehurst.wiremock.common.FileSource;
-import com.github.tomakehurst.wiremock.common.SingleRootFileSource;
-import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
-import com.github.tomakehurst.wiremock.extension.Parameters;
-import com.github.tomakehurst.wiremock.extension.ResponseTransformer;
-import com.github.tomakehurst.wiremock.http.Request;
-import com.github.tomakehurst.wiremock.http.Response;
-import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import com.github.tomakehurst.wiremock.stubbing.Scenario;
-import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.EnumSet;
-import java.util.HashSet;
 import java.util.Iterator;
-import java.util.NoSuchElementException;
 import jenkins.scm.api.SCMHeadObserver;
-import jenkins.scm.api.mixin.ChangeRequestCheckoutStrategy;
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
-import org.jvnet.hudson.test.Issue;
-import org.jvnet.hudson.test.JenkinsRule;
 import org.kohsuke.github.GHFileNotFoundException;
-import org.kohsuke.github.GHException;
-import org.kohsuke.github.GHRef;
 import org.kohsuke.github.GHRepository;
-import org.kohsuke.github.GitHub;
 import org.kohsuke.github.GHBranch;
-import org.kohsuke.github.PagedIterator;
-import org.kohsuke.github.PagedIterable;
 import org.mockito.Mockito;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.*;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.fail;
 
-public class GithubSCMSourceBranchesTest {
-
-    /**
-     * All tests in this class only use Jenkins for the extensions
-     */
-    @ClassRule
-    public static JenkinsRule r = new JenkinsRule();
-
-    public static WireMockRuleFactory factory = new WireMockRuleFactory();
-
-    @Rule
-    public WireMockRule githubRaw = factory.getRule(WireMockConfiguration.options()
-            .dynamicPort()
-            .usingFilesUnderClasspath("raw")
-    );
-    @Rule
-    public WireMockRule githubApi = factory.getRule(WireMockConfiguration.options()
-            .dynamicPort()
-            .usingFilesUnderClasspath("api")
-            .extensions(
-                    new ResponseTransformer() {
-                        @Override
-                        public Response transform(Request request, Response response, FileSource files,
-                                                  Parameters parameters) {
-                            if ("application/json"
-                                    .equals(response.getHeaders().getContentTypeHeader().mimeTypePart())) {
-                                return Response.Builder.like(response)
-                                        .but()
-                                        .body(response.getBodyAsString()
-                                                .replace("https://api.github.com/",
-                                                        "http://localhost:" + githubApi.port() + "/")
-                                                .replace("https://raw.githubusercontent.com/",
-                                                        "http://localhost:" + githubRaw.port() + "/")
-                                        )
-                                        .build();
-                            }
-                            return response;
-                        }
-
-                        @Override
-                        public String getName() {
-                            return "url-rewrite";
-                        }
-
-                    })
-    );
-    private GitHubSCMSource source;
-    private GitHub github;
-    private GHRepository repo;
+public class GithubSCMSourceBranchesTest extends GitSCMSourceBase {
 
     public GithubSCMSourceBranchesTest() {
         this.source = new GitHubSCMSource("cloudbeers", "yolo", null, false);
     }
-
-    @Before
-    public void prepareMockGitHub() throws Exception {
-        githubApi.stubFor(
-                get(urlMatching(".*")).atPriority(10).willReturn(aResponse().proxiedFrom("https://api.github.com/")));
-        githubRaw.stubFor(get(urlMatching(".*")).atPriority(10)
-                .willReturn(aResponse().proxiedFrom("https://raw.githubusercontent.com/")));
-
-        if (source.isConfiguredByUrl()) {
-            source = new GitHubSCMSource("cloudbeers", "yolo", "http://localhost:" + githubApi.port() + "/cloudbeers/yolo", true);
-        } else {
-            source.setApiUri("http://localhost:" + githubApi.port());
-        }
-        source.setTraits(Arrays.asList(new BranchDiscoveryTrait(true, true), new ForkPullRequestDiscoveryTrait(EnumSet.of(ChangeRequestCheckoutStrategy.MERGE), new ForkPullRequestDiscoveryTrait.TrustContributors())));
-        github = Connector.connect("http://localhost:" + githubApi.port(), null);
-        repo = github.getRepository("cloudbeers/yolo");
-    }
-
 
     @Test
     public void testMissingSingleBranch () throws IOException {
@@ -125,6 +35,7 @@ public class GithubSCMSourceBranchesTest {
                                         .withStatus(404)
                                         .withHeader("Content-Type", "application/json; charset=utf-8")
                                         .withBodyFile("../branches/_files/body-yolo-branches-non-existent-branch.json")));
+        //stubFor($TYPE(branch/PR/tag), $STATUS, $SCENARIO_NAME)
         SCMHeadObserver mockSCMHeadObserver = Mockito.mock(SCMHeadObserver.class);
         Mockito.when(mockSCMHeadObserver.getIncludes()).thenReturn(Collections
                 .singleton(new BranchSCMHead("non-existent-branch")));
