+++ src/test/java/org/jenkinsci/plugins/github_branch_source/GithubSCMSourceTagsTest.java
@@ -1,117 +1,34 @@
 package org.jenkinsci.plugins.github_branch_source;
 
-import com.github.tomakehurst.wiremock.common.FileSource;
-import com.github.tomakehurst.wiremock.common.SingleRootFileSource;
-import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
-import com.github.tomakehurst.wiremock.extension.Parameters;
-import com.github.tomakehurst.wiremock.extension.ResponseTransformer;
-import com.github.tomakehurst.wiremock.http.Request;
-import com.github.tomakehurst.wiremock.http.Response;
-import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import jenkins.scm.api.SCMHeadObserver;
-import jenkins.scm.api.mixin.ChangeRequestCheckoutStrategy;
-import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.EnumSet;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.NoSuchElementException;
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Rule;
 import org.junit.Test;
 import org.jvnet.hudson.test.Issue;
-import org.jvnet.hudson.test.JenkinsRule;
 import org.kohsuke.github.GHFileNotFoundException;
 import org.kohsuke.github.GHException;
 import org.kohsuke.github.GHRef;
 import org.kohsuke.github.GHRepository;
-import org.kohsuke.github.GitHub;
 import org.kohsuke.github.PagedIterator;
 import org.kohsuke.github.PagedIterable;
 import org.mockito.Mockito;
 
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.fail;
+import static org.junit.Assert.assertNotEquals;
 
-import static com.github.tomakehurst.wiremock.client.WireMock.*;
-import static org.junit.Assert.*;
 
-public class GithubSCMSourceTagsTest {
-
-    /**
-     * All tests in this class only use Jenkins for the extensions
-     */
-    @ClassRule
-    public static JenkinsRule r = new JenkinsRule();
-
-    public static WireMockRuleFactory factory = new WireMockRuleFactory();
-
-    @Rule
-    public WireMockRule githubRaw = factory.getRule(WireMockConfiguration.options()
-            .dynamicPort()
-            .usingFilesUnderClasspath("raw")
-    );
-    @Rule
-    public WireMockRule githubApi = factory.getRule(WireMockConfiguration.options()
-            .dynamicPort()
-            .usingFilesUnderClasspath("api")
-            .extensions(
-                    new ResponseTransformer() {
-                        @Override
-                        public Response transform(Request request, Response response, FileSource files,
-                                                  Parameters parameters) {
-                            if ("application/json"
-                                    .equals(response.getHeaders().getContentTypeHeader().mimeTypePart())) {
-                                return Response.Builder.like(response)
-                                        .but()
-                                        .body(response.getBodyAsString()
-                                                .replace("https://api.github.com/",
-                                                        "http://localhost:" + githubApi.port() + "/")
-                                                .replace("https://raw.githubusercontent.com/",
-                                                        "http://localhost:" + githubRaw.port() + "/")
-                                        )
-                                        .build();
-                            }
-                            return response;
-                        }
-
-                        @Override
-                        public String getName() {
-                            return "url-rewrite";
-                        }
-
-                    })
-    );
-    private GitHubSCMSource source;
-    private GitHub github;
-    private GHRepository repo;
+public class GithubSCMSourceTagsTest extends GitSCMSourceBase {
 
     public GithubSCMSourceTagsTest() {
         this.source = new GitHubSCMSource("cloudbeers", "yolo", null, false);
-    }
-    
-
-    @Before
-    public void prepareMockGitHub() throws Exception {
-        new File("src/test/resources/api/mappings").mkdirs();
-        new File("src/test/resources/api/__files").mkdirs();
-        new File("src/test/resources/raw/mappings").mkdirs();
-        new File("src/test/resources/raw/__files").mkdirs();
-        githubApi.enableRecordMappings(new SingleRootFileSource("src/test/resources/api/mappings"),
-                new SingleRootFileSource("src/test/resources/api/__files"));
-        githubRaw.enableRecordMappings(new SingleRootFileSource("src/test/resources/raw/mappings"),
-                new SingleRootFileSource("src/test/resources/raw/__files"));
-
-        githubApi.stubFor(
-                get(urlMatching(".*")).atPriority(10).willReturn(aResponse().proxiedFrom("https://api.github.com/")));
-        githubRaw.stubFor(get(urlMatching(".*")).atPriority(10)
-                .willReturn(aResponse().proxiedFrom("https://raw.githubusercontent.com/")));
-        source.setApiUri("http://localhost:" + githubApi.port());
-        source.setTraits(Arrays.asList(new BranchDiscoveryTrait(true, true), new ForkPullRequestDiscoveryTrait(EnumSet.of(ChangeRequestCheckoutStrategy.MERGE), new ForkPullRequestDiscoveryTrait.TrustContributors())));
-        github = Connector.connect("http://localhost:" + githubApi.port(), null);
-        repo = github.getRepository("cloudbeers/yolo");
     }
 
     @Test
@@ -148,6 +65,19 @@ public class GithubSCMSourceTagsTest {
         assertTrue(tags.hasNext());
         assertEquals("refs/tags/existent-tag", tags.next().getRef());
         assertFalse(tags.hasNext());
+    }
+
+    @Test
+    public void testRemoveNotAllowed() throws IOException {
+        SCMHeadObserver mockSCMHeadObserver = Mockito.mock(SCMHeadObserver.class);
+
+        Mockito.when(mockSCMHeadObserver.getIncludes()).thenReturn(Collections
+                .singleton(new GitHubTagSCMHead("existent-tag", System.currentTimeMillis())));
+        GitHubSCMSourceContext context = new GitHubSCMSourceContext(null, mockSCMHeadObserver);
+        context.wantTags(true);
+        GitHubSCMSourceRequest request = context.newRequest(new GitHubSCMSource("cloudbeers", "yolo", null, false), null);
+        Iterator<GHRef>  tags = new GitHubSCMSource.LazyTags(request, repo).iterator();
+        assertTrue(tags.hasNext());
     }
 
     @Test
@@ -283,7 +213,24 @@ public class GithubSCMSourceTagsTest {
 
         PagedIterator<GHRef> iteratorSpy = Mockito.spy(iterableSpy.iterator());
         Mockito.when(iterableSpy.iterator()).thenReturn(iteratorSpy);
-        Mockito.when(iteratorSpy.hasNext()).thenThrow(expectedError).thenThrow(expectedError2);
+        Mockito.when(iteratorSpy.hasNext()).thenThrow(expectedError);
+        Iterator<GHRef>  tags = new GitHubSCMSource.LazyTags(request, repoSpy).iterator();
+        assertFalse(tags.hasNext());
+    }
+    @Test
+    public void testExistingMultipleTagsIteratorIOErrorOnHasNext() throws IOException {
+        SCMHeadObserver mockSCMHeadObserver = Mockito.mock(SCMHeadObserver.class);
+        Error expectedError = new Error("Bad Tag Request", new IOException());
+        Mockito.when(mockSCMHeadObserver.getIncludes()).thenReturn(
+                new HashSet<>(Arrays.asList(
+                        new GitHubTagSCMHead("existent-multiple-tags1", System.currentTimeMillis()),
+                        new GitHubTagSCMHead("existent-multiple-tags2", System.currentTimeMillis()))));
+        GitHubSCMSourceContext context = new GitHubSCMSourceContext(null, mockSCMHeadObserver);
+        context.wantTags(true);
+        GitHubSCMSourceRequest request = context.newRequest(new GitHubSCMSource("cloudbeers", "yolo", null, false), null);
+        GHRepository repoSpy = Mockito.spy(repo);
+        PagedIterable<GHRef> iterableSpy = Mockito.spy(repoSpy.listRefs("tags"));
+        Mockito.when(repoSpy.listRefs("tags")).thenReturn(iterableSpy);
 
         //Expected: When initially getting multiple tags, return a filenotfound on hasNext which means it will get an empty list
         //Then return a IO error on the second hasNext
